modularize the C++ build
